# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-офис собственника",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/tasks": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "operationId": "create_task_api_v1_tasks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128,
                  "pattern": "^[A-Za-z0-9_-]+$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "office"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_api_v1_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  },
                  "title": "Response List Tasks Api V1 Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/daily-cash-and-receivable-risk": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Run Daily Cash And Receivable Risk",
        "description": "Run the first repeatable owner control; it only reads synthetic accounting data.",
        "operationId": "run_daily_cash_and_receivable_risk_api_v1_tasks_daily_cash_and_receivable_risk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyControlCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/process-graph": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Process Graph",
        "operationId": "get_process_graph_api_v1_tasks__task_id__process_graph_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessGraph"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/cancel": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Cancel Task",
        "operationId": "cancel_task_api_v1_tasks__task_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/clarify": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Clarify Task",
        "operationId": "clarify_task_api_v1_tasks__task_id__clarify_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegalClarification"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/retry": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Retry Task",
        "operationId": "retry_task_api_v1_tasks__task_id__retry_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128,
                  "pattern": "^[A-Za-z0-9_-]+$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentResult": {
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/AgentType"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Actions"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "agent",
          "status",
          "summary"
        ],
        "title": "AgentResult"
      },
      "AgentType": {
        "type": "string",
        "enum": [
          "orchestrator",
          "accountant",
          "lawyer",
          "security"
        ],
        "title": "AgentType"
      },
...
      "DailyControlCreate": {
        "properties": {
          "effective_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective On"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DailyControlCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LegalClarification": {
        "properties": {
          "jurisdiction": {
            "type": "string",
            "pattern": "^[A-Z]{2}$",
            "title": "Jurisdiction"
          },
          "effective_on": {
            "type": "string",
            "format": "date",
            "title": "Effective On"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "jurisdiction",
          "effective_on"
        ],
        "title": "LegalClarification"
      },
      "LegalResult": {
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/AgentType"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Actions"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "default": false
          },
          "mode": {
            "type": "string",
            "enum": [
              "legal_retrieval_pilot",
              "legal_analysis_pilot"
            ],
            "title": "Mode",
            "default": "legal_retrieval_pilot"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction"
          },
          "effective_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective On"
          },
          "legal_sources_found": {
            "type": "boolean",
            "title": "Legal Sources Found",
            "default": false
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "type": "array",
            "title": "Evidence"
          },
          "unresolved_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unresolved Questions"
          },
          "specialist_review_recommended": {
            "type": "boolean",
            "title": "Specialist Review Recommended",
            "default": true
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentExcerpt"
            },
            "type": "array",
            "title": "Documents"
          },
          "verified_references": {
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "type": "array",
            "title": "Verified References"
          },
          "analysis_model": {
            "type": "string",
            "title": "Analysis Model",
            "default": "none-retrieval-only"
          },
          "analysis_attempted": {
            "type": "boolean",
            "title": "Analysis Attempted",
            "default": false
          },
          "model_calls": {
            "items": {
              "$ref": "#/components/schemas/ModelCallInfo"
            },
            "type": "array",
            "title": "Model Calls"
          },
          "evidence_assessment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceAssessment"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence_assessment_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Assessment Version"
          },
          "consultant_plus": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConsultantPlusTrail"
              },
              {
                "type": "null"
              }
            ]
          },
          "retrieval_version": {
            "type": "string",
            "title": "Retrieval Version",
            "default": "mock-hash-v1"
          }
        },
        "type": "object",
        "required": [
          "agent",
          "status",
          "summary",
          "jurisdiction",
          "effective_on"
        ],
        "title": "LegalResult"
      },
...
      "OfficeResult": {
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/AgentType"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Actions"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "default": false
          },
          "mode": {
            "type": "string",
            "const": "office_mock_v1",
            "title": "Mode",
            "default": "office_mock_v1"
          },
          "accounting": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountingResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "legal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LegalResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "security": {
            "$ref": "#/components/schemas/SecurityResult"
          },
          "policy_decision": {
            "$ref": "#/components/schemas/PolicyDecision"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version",
            "default": "tz-v2.1-policy-v1"
          },
          "policy_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Policy Reasons"
          },
          "approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "draft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Id"
          }
        },
        "type": "object",
        "required": [
          "agent",
          "status",
          "summary",
          "security",
          "policy_decision"
        ],
        "title": "OfficeResult"
      },
...
      "ProcessGraph": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "process_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Process Id"
          },
          "process_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Process Version"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/ProcessGraphNode"
            },
            "type": "array",
            "title": "Nodes"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "nodes"
        ],
        "title": "ProcessGraph"
      },
...
      "RiskLevel": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "RiskLevel"
      },
      "RunBudget": {
        "properties": {
          "max_steps": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "default": 8
          },
          "max_llm_calls": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Max Llm Calls",
            "default": 2
          },
          "max_rework_cycles": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Max Rework Cycles",
            "default": 2
          },
          "max_runtime_seconds": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Max Runtime Seconds",
            "default": 120
          },
          "max_input_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Input Tokens",
            "default": 100000
          },
          "max_output_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Output Tokens",
            "default": 20000
          },
          "max_cost_rub": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Max Cost Rub",
            "default": "100.00"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RunBudget",
        "description": "Hard, persisted limits for a single process run."
      },
...
      "TaskCategory": {
        "type": "string",
        "enum": [
          "accounting",
          "legal",
          "security",
          "mixed",
          "unsupported"
        ],
        "title": "TaskCategory"
      },
      "TaskCreate": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Message"
          },
          "attachment_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Attachment Ids"
          },
          "requested_agent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentType"
              },
              {
                "type": "null"
              }
            ]
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z]{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction"
          },
          "effective_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective On"
          },
          "invoice_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^inv_[0-9]{3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Id"
          },
          "requested_action": {
            "type": "string",
            "enum": [
              "analyze",
              "prepare_payment_draft"
            ],
            "title": "Requested Action",
            "default": "analyze"
          },
          "process_id": {
            "type": "string",
            "enum": [
              "office_review",
              "daily_cash_and_receivable_risk"
            ],
            "title": "Process Id",
            "default": "office_review"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "TaskCreate"
      },
      "TaskResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "state": {
            "$ref": "#/components/schemas/TaskState"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OfficeResult"
              },
              {
                "$ref": "#/components/schemas/LegalResult"
              },
              {
                "$ref": "#/components/schemas/DailyOwnerDigest"
              },
              {
                "$ref": "#/components/schemas/AgentResult"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "process": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProcessDefinition"
              },
              {
                "$ref": "#/components/schemas/ProcessDefinitionV2"
              },
              {
                "$ref": "#/components/schemas/ProcessDefinitionV3"
              },
              {
                "type": "null"
              }
            ],
            "title": "Process"
          },
          "run_budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunBudget"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunUsage"
              },
              {
                "type": "null"
              }
            ]
          },
          "rework_parent_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rework Parent Task Id"
          },
          "rework_target_node_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rework Target Node Ids"
          },
          "rework_cycle": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Rework Cycle",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "state",
          "created_at",
          "updated_at"
        ],
        "title": "TaskResponse"
      },
      "TaskState": {
        "type": "string",
        "enum": [
          "draft",
          "queued",
          "classifying",
          "planning",
          "running",
          "waiting_input",
          "waiting_source",
          "waiting_approval",
          "completed",
          "failed",
          "failed_safe",
          "cancelled"
        ],
        "title": "TaskState"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}